+++ examples/telnet.js
@@ -1,4 +1,5 @@
 'use strict';
+
 var
   config = {
     'port': 3000,